chore: moved bloat related things to bloat.rs
