+++ crates/dry_run_core/src/schema/mod.rs
@@ -6,6 +6,7 @@ mod snapshot;
 mod types;
 pub mod vacuum;
 
+pub use bloat::*;
 pub use hash::{HashInput, compute_content_hash};
 pub use introspect::{
     fetch_is_standby, introspect_activity_stats, introspect_planner_stats, introspect_schema,
